change approach, fix tests
