+++ mlflow/store/model_registry/sqlalchemy_store.py
@@ -1098,3 +1098,6 @@ class SqlAlchemyStore(AbstractStore):
                 raise MlflowException(
                     f"Registered model alias {alias} not found.", INVALID_PARAMETER_VALUE
                 )
+
+    def _await_model_version_creation(self, mv, await_creation_for):
+        pass
